Rename application-server.[ch] to application-server-context.[ch]
