+++ src/5gmsaf/application-server-context.c
@@ -8,7 +8,7 @@ program. If this file is missing then the license can be retrieved from
 https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
 */
 
-#include "application-server.h"
+#include "application-server-context.h"
 #include "context.h"
 #include "utilities.h"
