+++ src/5gmsaf/provisioning-session.c
@@ -9,7 +9,7 @@ https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
 */
 
 #include "provisioning-session.h"
-#include "application-server.h"
+#include "application-server-context.h"
 #include "media-player-entry.h"
 #include "context.h"
 #include "utilities.h"
