+++ src/5gmsaf/service-access-information.h
@@ -23,7 +23,7 @@ https://drive.google.com/file/d/1cinCiA778IErENZ3JN52VFW-1ffHpx7Z/view
 #include <stdlib.h>
 #include "openapi/model/service_access_information_resource.h"
 #include "provisioning-session.h"
-#include "application-server.h"
+#include "application-server-context.h"
 
 #ifdef __cplusplus
 extern "C" {
